requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
Fabric==1.8.3
ecdsa==0.11
oauthlib==0.6.1
paramiko==1.12.3
pycrypto==2.6.1
python-twitter==1.3.1
requests==2.2.1
requests-oauthlib==0.4.0
simplejson==3.3.3
wsgiref==0.1.2
